+++ lib/src/native/realm_core.dart
@@ -130,7 +130,7 @@ class _RealmCore {
 
           if (schemaProperty.primaryKey) {
             classInfo.primary_key = schemaProperty.name.toCharPtr(arena);
-            propInfo.flags = realm_property_flags.RLM_PROPERTY_PRIMARY_KEY;
+            propInfo.flags |= realm_property_flags.RLM_PROPERTY_PRIMARY_KEY;
           }
         }
 
@@ -531,7 +531,7 @@ class _RealmCore {
     return RealmObjectHandle._(realmPtr);
   }
 
-  RealmObjectHandle getOrCreateRealmObjectWithPrimaryKey(Realm realm, int classKey, Object primaryKey) {
+  RealmObjectHandle getOrCreateRealmObjectWithPrimaryKey(Realm realm, int classKey, Object? primaryKey) {
     return using((Arena arena) {
       final realm_value = _toRealmValue(primaryKey, arena);
       final didCreate = arena<Bool>();
@@ -545,7 +545,7 @@ class _RealmCore {
     });
   }
 
-  RealmObjectHandle createRealmObjectWithPrimaryKey(Realm realm, int classKey, Object primaryKey) {
+  RealmObjectHandle createRealmObjectWithPrimaryKey(Realm realm, int classKey, Object? primaryKey) {
     return using((Arena arena) {
       final realm_value = _toRealmValue(primaryKey, arena);
       final realmPtr = _realmLib.invokeGetPointer(() => _realmLib.realm_object_create_with_primary_key(realm.handle._pointer, classKey, realm_value.ref));
@@ -576,7 +576,7 @@ class _RealmCore {
   // ignore: unused_element
   int get _threadId => _realmLib.realm_dart_get_thread_id();
 
-  RealmObjectHandle? find(Realm realm, int classKey, Object primaryKey) {
+  RealmObjectHandle? find(Realm realm, int classKey, Object? primaryKey) {
     return using((Arena arena) {
       final realm_value = _toRealmValue(primaryKey, arena);
       final pointer = _realmLib.realm_object_find_with_primary_key(realm.handle._pointer, classKey, realm_value.ref, nullptr);