+++ test/realm_object_test.dart
@@ -26,13 +26,16 @@ import 'test.dart';
 
 part 'realm_object_test.g.dart';
 
-const int maxInt = 9223372036854775807;
-const int minInt = -9223372036854775808;
-
 @RealmModel()
 class _ObjectIdPrimaryKey {
   @PrimaryKey()
   late ObjectId id;
+}
+
+@RealmModel()
+class _NullableObjectIdPrimaryKey {
+  @PrimaryKey()
+  late ObjectId? id;
 }
 
 @RealmModel()
@@ -42,15 +45,33 @@ class _IntPrimaryKey {
 }
 
 @RealmModel()
+class _NullableIntPrimaryKey {
+  @PrimaryKey()
+  int? id;
+}
+
+@RealmModel()
 class _StringPrimaryKey {
   @PrimaryKey()
   late String id;
 }
 
 @RealmModel()
+class _NullableStringPrimaryKey {
+  @PrimaryKey()
+  late String? id;
+}
+
+@RealmModel()
 class _UuidPrimaryKey {
   @PrimaryKey()
   late Uuid id;
+}
+
+@RealmModel()
+class _NullableUuidPrimaryKey {
+  @PrimaryKey()
+  late Uuid? id;
 }
 
 @RealmModel()
@@ -296,7 +317,7 @@ Future<void> main([List<String>? args]) async {
     await Future<void>.delayed(Duration(milliseconds: 20));
   });
 
-  void testPrimaryKey<T extends RealmObject, K extends Object>(SchemaObject schema, T Function() createObject, K key) {
+  void testPrimaryKey<T extends RealmObject, K extends Object>(SchemaObject schema, T Function() createObject, K? key) {
     test("$T primary key: $key", () {
       final pkProp = schema.properties.where((p) => p.primaryKey).single;
       final realm = Realm(Configuration.local([schema]));
@@ -314,14 +335,22 @@ Future<void> main([List<String>? args]) async {
     });
   }
 
-  final ints = [1, 0, -1, maxInt, minInt];
+  final ints = [1, 0, -1, maxInt, jsMaxInt, minInt, jsMinInt];
   for (final pk in ints) {
     testPrimaryKey(IntPrimaryKey.schema, () => IntPrimaryKey(pk), pk);
+  }
+
+  for (final pk in [null, ...ints]) {
+    testPrimaryKey(NullableIntPrimaryKey.schema, () => NullableIntPrimaryKey(pk), pk);
   }
 
   final strings = ["", "1", "abc", "null"];
   for (final pk in strings) {
     testPrimaryKey(StringPrimaryKey.schema, () => StringPrimaryKey(pk), pk);
+  }
+
+  for (final pk in [null, ...strings]) {
+    testPrimaryKey(NullableStringPrimaryKey.schema, () => NullableStringPrimaryKey(pk), pk);
   }
 
   final objectIds = [
@@ -334,12 +363,21 @@ Future<void> main([List<String>? args]) async {
     testPrimaryKey(ObjectIdPrimaryKey.schema, () => ObjectIdPrimaryKey(pk), pk);
   }
 
+  for (final pk in [null, ...objectIds]) {
+    testPrimaryKey(NullableObjectIdPrimaryKey.schema, () => NullableObjectIdPrimaryKey(pk), pk);
+  }
+
   final uuids = [
     Uuid.fromString('0f1dea4d-074e-4c72-b505-e2e8a727602f'),
     Uuid.fromString('00000000-0000-0000-0000-000000000000'),
   ];
+
   for (final pk in uuids) {
     testPrimaryKey(UuidPrimaryKey.schema, () => UuidPrimaryKey(pk), pk);
+  }
+
+  for (final pk in [null, ...uuids]) {
+    testPrimaryKey(NullableUuidPrimaryKey.schema, () => NullableUuidPrimaryKey(pk), pk);
   }
 
   test('Remapped property has correct names in Core', () {