+++ test/realm_test.dart
@@ -19,7 +19,6 @@
 // ignore_for_file: unused_local_variable, avoid_relative_lib_imports
 
 import 'dart:io';
-import 'dart:math';
 import 'package:test/test.dart' hide test, throws;
 import '../lib/realm.dart';